Got the ServerConnector working
3 minutes of ChatGPT work followed by 4 days of human debugging, what a time saver!
I probably could have had ChatGPT do some of that debugging to be fair, and one of the biggest bugs was actually caused by one of my changes. I'm not sure that vibe coding is better than regular coding, but I'm also not convinced it's entirely worse... I might use it again in the future. This phenomenon merits further investigation.